feat: Validator에서 PatterUtils 분리
