+++ src/main/java/com/programmers/springbootbasic/common/util/Validator.java
@@ -2,7 +2,6 @@ package com.programmers.springbootbasic.common.util;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
-import java.util.regex.Pattern;
 
 public final class Validator {
     private static final String PERCENT = "정률 할인";
@@ -11,8 +10,6 @@ public final class Validator {
     private static final String NULL_OR_BLANK = "빈 값이나 공백을 사용할 수 없습니다. 현재 값: ";
     private static final String INVALID_VOUCHER_TYPE = "잘못된 바우처 유형입니다. 현재 입력 값: ";
     private static final String DATETIME_IS_NULL = "빈 값이나 공백을 날짜/시간으로 지정할 수 없습니다. 현재 값: ";
-    private static final Pattern DATE_PATTERN = Pattern.compile("^\\d{4}-\\d{2}-\\d{2}$");
-    private static final Pattern DATETIME_PATTERN = Pattern.compile("^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}\\.\\d{3}$\n");
 
     private Validator() {
     }
@@ -41,13 +38,4 @@ public final class Validator {
             throw new IllegalArgumentException(DATETIME_IS_NULL + input);
         }
     }
-
-    public static boolean isDate(String input) {
-        return DATE_PATTERN.matcher(input).matches();
-    }
-
-    public static boolean isDateTime(String input) {
-        return DATETIME_PATTERN.matcher(input).matches();
-    }
-
 }
